added utils folder and mocData and constant file
