+++ src/utils/constant.js
@@ -1,0 +1,2 @@
+const LOGO_URL="https://www.logodesign.net/logo/smoking-burger-with-lettuce-3624ld.png";
+const CDN_URL="https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/";
